+++ mne/utils.py
@@ -3207,6 +3207,8 @@ def _hid_match(event_id, keys):
 
 
 class GetEpochsMixin(object):
+    """Class to add epoch selection and metadata to certain classes."""
+
     def __getitem__(self, item):
         """Return an Epochs object with a copied subset of epochs.
 
@@ -3334,7 +3336,8 @@ class GetEpochsMixin(object):
         if drop_event_id:
             # update event id to reflect new content of inst
             inst.event_id = dict((k, v) for k, v in inst.event_id.items()
-                                   if v in inst.events[:, 2])
+                                 if v in inst.events[:, 2])
+
         if return_indices:
             return inst, select
         else:
@@ -3342,7 +3345,7 @@ class GetEpochsMixin(object):
 
     def _keys_to_idx(self, keys):
         """Find entries in event dict."""
-        keys = [keys] if not isinstance(keys, (list, tuple)) else keys
+        keys = keys if isinstance(keys, (list, tuple)) else [keys]
         try:
             # Assume it's a condition name
             return np.where(np.any(
@@ -3489,8 +3492,11 @@ class GetEpochsMixin(object):
                                      % (len(metadata), len(self.events)))
                 if reset_index:
                     if hasattr(self, 'selection'):
-                        metadata = metadata.reset_index(drop=True)  # makes a copy
+                        # makes a copy
+                        metadata = metadata.reset_index(drop=True)
                         metadata.index = self.selection
+                    else:
+                        metadata = deepcopy(metadata)
             else:
                 _validate_type(metadata, types=list,
                                item_name='metadata')